+++ src/Oro/Bundle/PricingBundle/ImportExport/Reader/PriceListProductPricesReader.php
@@ -2,13 +2,19 @@
 
 namespace Oro\Bundle\PricingBundle\ImportExport\Reader;
 
+use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\ORM\Query;
+use Doctrine\ORM\QueryBuilder;
 use Oro\Bundle\ImportExportBundle\Context\ContextInterface;
 use Oro\Bundle\ImportExportBundle\Reader\EntityReader;
 use Oro\Bundle\OrganizationBundle\Entity\Organization;
 use Oro\Bundle\PricingBundle\ORM\Walker\PriceShardWalker;
 use Oro\Bundle\PricingBundle\Sharding\ShardManager;
 
+/**
+ * Prepares the list of entities for export.
+ * Responsible for creating a list for each batch during export
+ */
 class PriceListProductPricesReader extends EntityReader
 {
     /**
@@ -68,5 +74,32 @@ class PriceListProductPricesReader extends EntityReader
     public function setShardManager(ShardManager $shardManager)
     {
         $this->shardManager = $shardManager;
+    }
+
+    /**
+     * @param ObjectManager $entityManager
+     * @param string $entityName
+     * @param array $options
+     *
+     * @return QueryBuilder
+     */
+    protected function createQueryBuilderByEntityNameAndIdentifier(
+        ObjectManager $entityManager,
+        string $entityName,
+        array $options = []
+    ): QueryBuilder {
+        if (!array_key_exists('price_list_id', $options)) {
+            throw new \LogicException('Unable to read prices, price_list_id should be defined.');
+        }
+
+        $queryBuilder = parent::createQueryBuilderByEntityNameAndIdentifier(
+            $entityManager,
+            $entityName,
+            $options
+        );
+        $queryBuilder->andWhere('o.priceList = :priceListId');
+        $queryBuilder->setParameter('priceListId', $options['price_list_id']);
+
+        return $queryBuilder;
     }
 }